Move some parts out of the main webpack config.
