+++ package.json
@@ -161,6 +161,7 @@
     "webpack-bundle-analyzer": "^2.11.2",
     "webpack-dev-middleware": "^3.1.3",
     "webpack-hot-middleware": "^2.22.1",
+    "webpack-merge": "^4.1.2",
     "webpack-pwa-manifest": "^3.6.2",
     "webpack-subresource-integrity": "^1.1.0-rc.4",
     "yaml-loader": "^0.5.0"
