+++ tasks/utils/htmlMinifierOptions.js
@@ -1,0 +1,13 @@
+// Minification options for html-minifier.
+module.exports = {
+  removeComments: true,
+  collapseWhitespace: true,
+  collapseBooleanAttributes: true,
+  removeTagWhitespace: true,
+  removeAttributeQuotes: true,
+  removeRedundantAttributes: true,
+  removeScriptTypeAttributes: true,
+  removeStyleLinkTypeAttributes: true,
+  removeOptionalTags: true,
+  minifyCSS: true,
+};
